requirements.txt
requests
PyYAML
python-slugify
pydantic
click
structlog
simhash
tenacity
python-dateutil
tqdm
langchain-anthropic
python-dotenv
pytest
